+++ CHANGELOG.md
@@ -27,5 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multi-Threaded WebSocket Listeners (per game)
 
 [Unreleased]: https://github.com/rshchekotov/wwbda/compare/v0.2.0...HEAD
+
 [0.2.0]: https://github.com/rshchekotov/wwbda/releases/tag/v0.2.0
+
 [0.1.0]: https://github.com/rshchekotov/wwbda/releases/tag/v0.1.0