+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2026-03-08
+## [Unreleased] - 2026-03-11
+
+## [0.3.0] - 2026-03-11
 
 ### Added
 
@@ -43,7 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rust-based LiShogi library
 - Multi-Threaded WebSocket Listeners (per game)
 
-[Unreleased]: https://github.com/rshchekotov/wwbda/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/rshchekotov/wwbda/compare/v0.3.0...HEAD
+
+[0.3.0]: https://github.com/rshchekotov/wwbda/releases/tag/v0.3.0
 
 [0.2.0]: https://github.com/rshchekotov/wwbda/releases/tag/v0.2.0